Fix blocking issue at test_tasks.py
