+++ tests/test_mm/test_tasks/test_mmtask.py
@@ -1,0 +1,37 @@
+from unittest.mock import patch
+
+import mmcls  # noqa: F401
+import mmdet  # noqa: F401
+import mmseg  # noqa: F401
+
+from siatune.mm.tasks import MMClassification, MMDetection, MMSegmentation
+
+
+@patch('mmcls.apis.train_model')
+@patch('mmcls.datasets.build_dataset')
+@patch('mmcls.models.build_classifier')
+def test_mmcls(*not_used):
+    task = MMClassification()
+    task_args = ['tests/data/config.py']
+    task.set_args(task_args)
+    task.run(args=task.args)
+
+
+@patch('mmdet.apis.train_detector')
+@patch('mmdet.datasets.build_dataset')
+@patch('mmdet.models.build_detector')
+def test_mmdet(*not_used):
+    task = MMDetection()
+    task_args = ['tests/data/config.py']
+    task.set_args(task_args)
+    task.run(args=task.args)
+
+
+@patch('mmseg.apis.train_segmentor')
+@patch('mmseg.datasets.build_dataset')
+@patch('mmseg.models.build_segmentor')
+def test_mmseg(*not_used):
+    task = MMSegmentation()
+    task_args = ['tests/data/config.py']
+    task.set_args(task_args)
+    task.run(args=task.args)
